pyproject.toml
[project]
name = "mcp-server-mariadb"
version = "0.1.2"
description = "MCP Server for MariaDB"
readme = "README.md"
authors = [
    { name = "HeejunShin", email = "abel9851@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "mariadb>=1.1.12",
    "mcp[cli]>=1.3.0",
    "mysql-connector-python>=9.2.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "sqlparse>=0.5.3",
]

[project.scripts]
mcp-server-mariadb = "mcp_server_mariadb:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
